+++ age/src/protocol.rs
@@ -19,6 +19,20 @@ use crate::{
 const HEADER_KEY_LABEL: &[u8] = b"header";
 const PAYLOAD_KEY_LABEL: &[u8] = b"payload";
 
+/// Callbacks that might be triggered during decryption.
+pub trait Callbacks {
+    /// Requests a passphrase to decrypt a key.
+    fn request_passphrase(&self, description: &str) -> Option<SecretString>;
+}
+
+struct NoCallbacks;
+
+impl Callbacks for NoCallbacks {
+    fn request_passphrase(&self, _description: &str) -> Option<SecretString> {
+        None
+    }
+}
+
 /// Handles the various types of age encryption.
 pub enum Encryptor {
     /// Encryption to a list of recipients identified by keys.
@@ -70,8 +84,13 @@ impl Encryptor {
 
 /// Handles the various types of age decryption.
 pub enum Decryptor {
-    /// Trial decryption against a list of secret keys.
-    Keys(Vec<Identity>),
+    /// Trial decryption against a list of identities.
+    Identities {
+        /// The identities to use.
+        identities: Vec<Identity>,
+        /// A handler for any callbacks triggered by an `Identity`.
+        callbacks: Box<dyn Callbacks>,
+    },
     /// Decryption with a passphrase.
     Passphrase {
         /// The passphrase to decrypt with.
@@ -83,6 +102,31 @@ pub enum Decryptor {
 }
 
 impl Decryptor {
+    /// Creates a decryptor with a list of identities.
+    ///
+    /// The decryptor will have no callbacks registered, so it will be unable to use
+    /// identities that require e.g. a passphrase to decrypt.
+    pub fn with_identities(identities: Vec<Identity>) -> Self {
+        Decryptor::Identities {
+            identities,
+            callbacks: Box::new(NoCallbacks),
+        }
+    }
+
+    /// Creates a decryptor with a list of identities and a callback handler.
+    ///
+    /// The decryptor will have no callbacks registered, so it will be unable to use
+    /// identities that require e.g. a passphrase to decrypt.
+    pub fn with_identities_and_callbacks(
+        identities: Vec<Identity>,
+        callbacks: Box<dyn Callbacks>,
+    ) -> Self {
+        Decryptor::Identities {
+            identities,
+            callbacks,
+        }
+    }
+
     /// Creates a decryptor with a passphrase and the default max work factor.
     pub fn with_passphrase(passphrase: SecretString) -> Self {
         Decryptor::Passphrase {
@@ -91,16 +135,20 @@ impl Decryptor {
         }
     }
 
-    fn unwrap_file_key<P: Fn(&str) -> Option<SecretString> + Copy>(
-        &self,
-        line: &RecipientLine,
-        request_passphrase: P,
-    ) -> Result<Option<FileKey>, Error> {
+    fn unwrap_file_key(&self, line: &RecipientLine) -> Result<Option<FileKey>, Error> {
         match (self, line) {
-            (Decryptor::Keys(_), RecipientLine::Scrypt(_)) => Err(Error::MessageRequiresPassphrase),
-            (Decryptor::Keys(keys), _) => keys
+            (Decryptor::Identities { .. }, RecipientLine::Scrypt(_)) => {
+                Err(Error::MessageRequiresPassphrase)
+            }
+            (
+                Decryptor::Identities {
+                    identities,
+                    callbacks,
+                },
+                _,
+            ) => identities
                 .iter()
-                .find_map(|key| key.unwrap_file_key(line, request_passphrase))
+                .find_map(|key| key.unwrap_file_key(line, callbacks.as_ref()))
                 .transpose(),
             (
                 Decryptor::Passphrase {
@@ -119,11 +167,7 @@ impl Decryptor {
     /// to be decrypted before it can be used to decrypt the message.
     ///
     /// If successful, returns a reader that will provide the plaintext.
-    pub fn trial_decrypt<R: Read, P: Fn(&str) -> Option<SecretString> + Copy>(
-        &self,
-        input: R,
-        request_passphrase: P,
-    ) -> Result<impl Read, Error> {
+    pub fn trial_decrypt<R: Read>(&self, input: R) -> Result<impl Read, Error> {
         let mut input = ArmoredReader::from_reader(input);
 
         match Header::read(&mut input)? {
@@ -135,21 +179,19 @@ impl Decryptor {
                     .recipients
                     .iter()
                     .find_map(|r| {
-                        self.unwrap_file_key(r, request_passphrase)
-                            .transpose()
-                            .map(|res| {
-                                res.and_then(|file_key| {
-                                    // Verify the MAC
-                                    header.verify_mac(hkdf(
-                                        &[],
-                                        HEADER_KEY_LABEL,
-                                        file_key.0.expose_secret(),
-                                    ))?;
+                        self.unwrap_file_key(r).transpose().map(|res| {
+                            res.and_then(|file_key| {
+                                // Verify the MAC
+                                header.verify_mac(hkdf(
+                                    &[],
+                                    HEADER_KEY_LABEL,
+                                    file_key.0.expose_secret(),
+                                ))?;
 
-                                    // Return the payload key
-                                    Ok(hkdf(&nonce, PAYLOAD_KEY_LABEL, file_key.0.expose_secret()))
-                                })
+                                // Return the payload key
+                                Ok(hkdf(&nonce, PAYLOAD_KEY_LABEL, file_key.0.expose_secret()))
                             })
+                        })
                     })
                     .unwrap_or(Err(Error::NoMatchingKeys))
                     .map(|payload_key| Stream::decrypt(&payload_key, input))
@@ -164,10 +206,9 @@ impl Decryptor {
     /// to be decrypted before it can be used to decrypt the message.
     ///
     /// If successful, returns a seekable reader that will provide the plaintext.
-    pub fn trial_decrypt_seekable<R: Read + Seek, P: Fn(&str) -> Option<SecretString> + Copy>(
+    pub fn trial_decrypt_seekable<R: Read + Seek>(
         &self,
         mut input: R,
-        request_passphrase: P,
     ) -> Result<StreamReader<R>, Error> {
         match Header::read(&mut input)? {
             Header::V1(header) => {
@@ -178,21 +219,19 @@ impl Decryptor {
                     .recipients
                     .iter()
                     .find_map(|r| {
-                        self.unwrap_file_key(r, request_passphrase)
-                            .transpose()
-                            .map(|res| {
-                                res.and_then(|file_key| {
-                                    // Verify the MAC
-                                    header.verify_mac(hkdf(
-                                        &[],
-                                        HEADER_KEY_LABEL,
-                                        file_key.0.expose_secret(),
-                                    ))?;
+                        self.unwrap_file_key(r).transpose().map(|res| {
+                            res.and_then(|file_key| {
+                                // Verify the MAC
+                                header.verify_mac(hkdf(
+                                    &[],
+                                    HEADER_KEY_LABEL,
+                                    file_key.0.expose_secret(),
+                                ))?;
 
-                                    // Return the payload key
-                                    Ok(hkdf(&nonce, PAYLOAD_KEY_LABEL, file_key.0.expose_secret()))
-                                })
+                                // Return the payload key
+                                Ok(hkdf(&nonce, PAYLOAD_KEY_LABEL, file_key.0.expose_secret()))
                             })
+                        })
                     })
                     .unwrap_or(Err(Error::NoMatchingKeys))
                     .and_then(|payload_key| {
@@ -229,8 +268,8 @@ mod tests {
             w.finish().unwrap();
         }
 
-        let d = Decryptor::Keys(sk);
-        let mut r = d.trial_decrypt(&encrypted[..], |_| None).unwrap();
+        let d = Decryptor::with_identities(sk);
+        let mut r = d.trial_decrypt(&encrypted[..]).unwrap();
         let mut decrypted = vec![];
         r.read_to_end(&mut decrypted).unwrap();
 
@@ -250,7 +289,7 @@ mod tests {
         }
 
         let d = Decryptor::with_passphrase(SecretString::new("passphrase".to_string()));
-        let mut r = d.trial_decrypt(&encrypted[..], |_| None).unwrap();
+        let mut r = d.trial_decrypt(&encrypted[..]).unwrap();
         let mut decrypted = vec![];
         r.read_to_end(&mut decrypted).unwrap();
 
@@ -274,8 +313,8 @@ mod tests {
             w.finish().unwrap();
         }
 
-        let d = Decryptor::Keys(sk);
-        let mut r = d.trial_decrypt(&encrypted[..], |_| None).unwrap();
+        let d = Decryptor::with_identities(sk);
+        let mut r = d.trial_decrypt(&encrypted[..]).unwrap();
         let mut decrypted = vec![];
         r.read_to_end(&mut decrypted).unwrap();
 
@@ -298,8 +337,8 @@ mod tests {
             w.finish().unwrap();
         }
 
-        let d = Decryptor::Keys(sk);
-        let mut r = d.trial_decrypt(&encrypted[..], |_| None).unwrap();
+        let d = Decryptor::with_identities(sk);
+        let mut r = d.trial_decrypt(&encrypted[..]).unwrap();
         let mut decrypted = vec![];
         r.read_to_end(&mut decrypted).unwrap();